# `ultralytics 8.1.43` 40% faster `ultralytics` imports