<!--
Thank you 🙏 for your contribution to [Ultralytics](https://ultralytics.com) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. **Check for Existing Contributions**: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. **Link Related Issues**: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. **Elaborate Your Changes**: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. **Ultralytics Contributor License Agreement (CLA)**: To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    _I have read the CLA Document and I sign the CLA_

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing). Your adherence to these guidelines ensures a faster and more effective review process.
--->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Optimizing Python and Torchvision Imports and Usage 🚀

### 📊 Key Changes
- **Simplification of import statements:** Transitioned from direct import of `ToTensor` to using `import torchvision.transforms as T`, streamlining the way torchvision transforms are accessed and applied.
- **Reduced redundant code:** Removed duplicate import statements for `torchvision.transforms` by refining the strategy for importing utilities across modules.
- **Introduced `PYTHON_VERSION` and `TORCHVISION_VERSION` tracking:** Centralized the handling of Python and Torchvision version checks, making it easier to manage compatibility and requirements across the project.
- **Minor refactoring:** Changes include tidying up code, enhancing readability, and ensuring consistency in how external libraries are utilized.

### 🎯 Purpose & Impact
- **Enhanced code maintainability:** By tidying up imports and centralizing version checks, the codebase becomes easier to navigate and maintain, reducing potential bugs or issues related to dependency management.
- **Improved efficiency:** Streamlining imports and removing redundant code can slightly enhance the performance of test suites and other parts of the codebase that rely on these components.
- **Smoother version management:** Having centralized checks for `PYTHON_VERSION` and `TORCHVISION_VERSION` aids in ensuring that the software is compatible with the environment it's running in, facilitating smoother updates and compatibility checks.
- **Overall user experience:** These backend changes, while mostly invisible to end-users, contribute to a more stable and reliable framework, indirectly benefiting users by reducing potential issues and improving the software's robustness.