@glenn-jocher 2024-04-04
@ddeevvaa I still haven't removed it completely, there's one file left, but seems to improve speed from about 1.4s -> 1.0s import