@glenn-jocher 2024-04-04
@ddeevvaa final results are here averaged over 30 imports. Speedup of 25.3% acheived, reduced import time from 1411ms by 357ms to 1054ms on my Macbook Air M2.

### this PR

- 1.054 s for `import ultralytics`

### `main` branch

- 1.411s for `import ultralytics`