@glenn-jocher 2024-04-05
@ddeevvaa thank you so much for the encouraging feedback! 🌟 We're thrilled by the performance improvements as well and can't wait for the merge to happen. It's all about making YOLOv8 better and faster for everyone. Keep an eye out for updates soon!